feat(mmc-image): fetch this repo on boot if it doesn’t exist
Software nodes are commonly developed on a system externally from the
Raspberry Pi. Problems arise when we want to test a node on the actual
hardware, because it would normally have to be packaged and built via
Nix, which would require us to reflash the MMC of the RPi, a
time-consuming operation (and certainly REPL). This commit adds a
convenient systemd unit that simply clones this repository to the file
system so that one may fetch a branch to the actual hardware and try
changes out, without requiring any system rebuilds.